requirements.txt
fastapi
uvicorn
temporalio
requests
kubernetes
strands-agents
boto3
pydantic